pyproject.toml
[project]
name = "airflow-datalake"
version = "0.1.0"
description = "Airflow Data Lake with Kafka, Spark, MinIO, StarRocks and Superset"
readme = "README-AIRFLOW.md"
requires-python = ">=3.10"
dependencies = [
    "pyspark>=4.0.0",
    "minio>=7.1.0",
    "kafka-python-ng>=2.2.2",
    "psycopg2-binary>=2.9.0",
    "redis>=4.5.0",
    "requests>=2.28.0",
    "pandas>=1.5.0",
    "pyarrow>=10.0.0",
    "streamlit>=1.49.1",
    "confluent-kafka>=2.11.1",
    "python-dotenv>=1.1.1",
    "grpcio>=1.74.0",
    "grpcio-status>=1.74.0",
    "apache-airflow-providers-postgres>=5.0.0",
    "apache-airflow-providers-redis>=3.0.0",
    "apache-airflow-providers-apache-spark>=4.0.0",
    "apache-airflow-providers-apache-kafka>=1.0.0",
]

[dependency-groups]
# Installed by default in dev with `uv sync` (see [tool.uv] below)
dev = [
    "pre-commit==4.2.0",
    "black",
    "pytest",
    "ruff",
    "isort",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
# Include your package files (adjust paths as needed)
include = ["*.py", "src/**/*.py"]

# ----------------------------
# Tool configuration
# ----------------------------

[tool.uv]
# Make `uv sync` include the "dev" group by default in your devcontainer
default-groups = ["dev"]

[tool.ruff]
# Lint + (optionally) format with Ruff
line-length = 140
target-version = "py310"
# Enable Pyflakes + pycodestyle subset + import sorting
select = ["E", "F", "I"]
ignore = ["F821", "E402"]
exclude = [".git", "__pycache__", "build", "dist", "setup.py"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
